move file column to API
